==> 02. VGG16 Model Training Part 1 - Transfer Learning.ipynb <==
GLOBAL_SEED = 7532

from numpy.random import seed
seed(GLOBAL_SEED)
from tensorflow import set_random_seed
set_random_seed(GLOBAL_SEED)

import numpy as np
import pandas as pd

from keras.models import Model, Sequential, load_model
from keras.callbacks import Callback, EarlyStopping, ModelCheckpoint, ReduceLROnPlateau
from keras.optimizers import Adam
from keras.layers import BatchNormalization, Dense, Dropout, Conv2D, Flatten, MaxPool2D
from keras.utils import to_categorical
from keras.preprocessing.image import ImageDataGenerator
from keras.applications import VGG16

from sklearn.metrics import f1_score
from sklearn.utils.class_weight import compute_class_weight

Read in the metadata.

train_set = pd.read_csv('train_set_metadata.csv')
valid_set = pd.read_csv('valid_set_metadata.csv')
test_set = pd.read_csv('test_set_metadata.csv')

Set the main constants.

INPUT_SHAPE = (256, 384, 3)

LEARNING_RATE = 0.0001
N_EPOCHS = 100
BATCH_SIZE = 32

Read in the VGG16 model without the top part.

vgg16_model = VGG16(weights='imagenet', include_top=False, input_shape=INPUT_SHAPE)
vgg16_model.summary()

Freeze the convolutional layers.

for layer in vgg16_model.layers:
    layer.trainable = False

Create the model top.

def create_model_top(input_shape):
    model = Sequential()
    
    dropout_rate = 0.25
    
    model.add(Flatten())
    model.add(Dropout(rate=dropout_rate, seed=GLOBAL_SEED))
    model.add(Dense(units=1024, activation='relu'))
    model.add(Dropout(rate=dropout_rate, seed=GLOBAL_SEED))
    model.add(Dense(units=256, activation='relu'))
    model.add(Dropout(rate=dropout_rate, seed=GLOBAL_SEED))
    model.add(Dense(5, activation='softmax'))
    
    return model

Attach the model top to the VGG16 model.

model_top = create_model_top(vgg16_model.outputs[0].get_shape().as_list()[1:])
model = Model(inputs=vgg16_model.inputs, outputs=model_top(vgg16_model.outputs[0]))
model.summary()

Verify that the appropriate layers are frozen.

for layer in model.layers:
    print(layer.trainable)

Create the f1 metric to be used during model training.

class F1_Metric(Callback):
    def on_train_begin(self, logs={}):
        self.f1_scores = []

    def on_epoch_end(self, epoch, logs={}):
        predict = np.round(np.asarray(self.model.predict(self.validation_data[0])))
        targ = self.validation_data[1]

        self.f1_scores.append(f1_score(targ, predict, average='weighted'))
        
        print(f' val_f1: {self.f1_scores[-1]}')
        
        return
    
f1_metric = F1_Metric()

Compute the class weights to be used during model training in order to fight the class imbalances.

class_values = train_set['category'].values - 1

classes = np.unique(class_values)
weights = compute_class_weight('balanced', np.unique(class_values), class_values)

class_weights = dict(zip(classes, weights))
class_weights

Train the model utilizing Adam optimizer, learning rate reduction on plateau, class weights and data augmentation.

# Prepare the training and validation data
X_train = np.load('train_set_hmgd_arr_256_384_VGG16.npy')
y_train = to_categorical(train_set['category'].values - 1)

X_valid = np.load('valid_set_hmgd_arr_256_384_VGG16.npy')
y_valid = to_categorical(valid_set['category'].values - 1)


# Create and compile the model
adam = Adam(lr=LEARNING_RATE)
model.compile(optimizer=adam, loss='categorical_crossentropy')


# Initialize callbacks
checkpoint = ModelCheckpoint('model_006_epoch_{epoch:03d}.hdf5', monitor='val_loss', 
                             save_best_only=False, save_weights_only=False)
lr_reduction = ReduceLROnPlateau(monitor='val_loss', factor=0.5, patience=6)
#early_stopping = EarlyStopping(monitor='val_loss', patience=12)    
callback_list = [f1_metric, checkpoint, lr_reduction]#, early_stopping]


# data augmentation
data_gen = ImageDataGenerator(rotation_range=20, 
                              width_shift_range=0.2, 
                              height_shift_range=0.2, 
                              horizontal_flip=True)

steps_per_epoch = int(len(X_train)) / BATCH_SIZE

model.fit_generator(data_gen.flow(X_train, y_train, batch_size=BATCH_SIZE), 
                    steps_per_epoch=steps_per_epoch, 
                    epochs=N_EPOCHS,
                    callbacks=callback_list,
                    validation_data=(X_valid, y_valid),
                    class_weight=class_weights,
                    workers=4, 
                    verbose=2)

# free up memory
del X_train
del X_valid

Read in the test data as well as the best model with respect to the f1 score.

X_test = np.load('test_set_hmgd_arr_256_384_VGG16.npy')
model = load_model('model_006_epoch_088.hdf5')

Compute the predictions and prepare the submission.

test_pred_classes = model.predict(X_test).argmax(axis=-1) + 1
test_pred_classes

sample_submission = pd.read_csv('data/sample_submission.csv')
submission = sample_submission.copy()

submission['category'] = test_pred_classes
submission['category'].value_counts()

submission.head()

submission.to_csv('model_006_submission_01.csv', index=False)

#                      Epoch 88
# Validation score:  0.9094737873
# Leaderboard score: 0.9155052841
