==> 03. Model Ensemble.ipynb <==
import numpy as np
import pandas as pd

from keras.models import load_model

Read in the data.

X_test = np.load('test_set_hmgd_arr_256_384_VGG16.npy')
submission = pd.read_csv('data/sample_submission.csv')

Read in the best 2 models with respect to the f1 score.

model_01 = load_model('model_010_FT001_epoch_023.hdf5')
model_02 = load_model('model_010_FT001_epoch_046.hdf5')

Compute the ensemble predictions and prepare the submission.

pred_prob_01 = model_01.predict(X_test)
pred_prob_02 = model_02.predict(X_test)

pred_prob_ensemble = (pred_prob_01 + pred_prob_02) / 2
pred_classes_ensemble = pred_prob_ensemble.argmax(axis=-1) + 1

submission['category'] = pred_classes_ensemble
submission['category'].value_counts()

submission.to_csv('model_010_ensemble_submission_01.csv', index=False)

This simple ensemble improves upon the score of its constituents. 

#                      Epoch 23      Epoch 46      Ensemble
# Validation score:  0.9578207955  0.9577930088
# Leaderboard score: 0.9617234462  0.9609120265  0.96272086496
